+++ src/SlaegtsAssistent.Core/Domain/Person.cs
@@ -35,4 +35,8 @@ public sealed class Person
     public IList<Source> Sources { get; } = new List<Source>();
 
     public IList<Media> Media { get; } = new List<Media>();
+
+    public IList<GedcomEvent> Events { get; } = new List<GedcomEvent>();
+
+    public IList<Census> Census { get; } = new List<Census>();
 }